Rename proofs -> proof
