+++ packages/api/src/beacon/routes/proof.ts
