+++ packages/api/test/unit/beacon/genericServerTest/proofs.test.ts
@@ -1,7 +1,7 @@
 import {config} from "@lodestar/config/default";
-import {Api, ReqTypes} from "../../../../src/beacon/routes/proofs.js";
+import {Api, ReqTypes} from "../../../../src/beacon/routes/proof.js";
 import {getClient} from "../../../../src/beacon/client/proofs.js";
-import {getRoutes} from "../../../../src/beacon/server/proofs.js";
+import {getRoutes} from "../../../../src/beacon/server/proof.js";
 import {runGenericServerTest} from "../../../utils/genericServerTest.js";
 import {testData} from "../testData/proofs.js";
