+++ packages/beacon-node/src/api/impl/api.ts
@@ -8,7 +8,7 @@ import {getEventsApi} from "./events/index.js";
 import {getLightclientApi} from "./lightclient/index.js";
 import {getLodestarApi} from "./lodestar/index.js";
 import {getNodeApi} from "./node/index.js";
-import {getProofsApi} from "./proofs/index.js";
+import {getProofApi} from "./proof/index.js";
 import {getValidatorApi} from "./validator/index.js";
 
 export function getApi(opts: IApiOptions, modules: ApiModules): Api {
@@ -20,7 +20,7 @@ export function getApi(opts: IApiOptions, modules: ApiModules): Api {
     lightclient: getLightclientApi(modules),
     lodestar: getLodestarApi(modules),
     node: getNodeApi(opts, modules),
-    proofs: getProofsApi(opts, modules),
+    proof: getProofApi(opts, modules),
     validator: getValidatorApi(modules),
   };
 }
