+++ packages/beacon-node/src/api/impl/proof/index.ts
@@ -4,10 +4,10 @@ import {ApiModules} from "../types.js";
 import {resolveStateId} from "../beacon/state/utils.js";
 import {IApiOptions} from "../../options.js";
 
-export function getProofsApi(
+export function getProofApi(
   opts: IApiOptions,
   {chain, config, db}: Pick<ApiModules, "chain" | "config" | "db">
-): routes.proofs.Api {
+): routes.proof.Api {
   // It's currently possible to request gigantic proofs (eg: a proof of the entire beacon state)
   // We want some some sort of resistance against this DoS vector.
   const maxGindicesInProof = opts.maxGindicesInProof ?? 512;
