+++ packages/validator/test/utils/apiStub.ts
@@ -18,7 +18,7 @@ export function getApiClientStub(
     lightclient: sandbox.stub(api.lightclient),
     lodestar: sandbox.stub(api.lodestar),
     node: sandbox.stub(api.node),
-    proofs: sandbox.stub(api.proofs),
+    proof: sandbox.stub(api.proof),
     validator: sandbox.stub(api.validator),
   };
 }
